adding more size to the cards
