+++ apps/booking-app/pages/admin/index.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import { LoginContainer } from '../../../../packages/admin-panel/src/components/LoginContainer'
 import { FiltersContainer } from '@booking-ui/admin-panel';
+import { ContextOfFilters } from '@booking-ui/admin-panel/src/components/Filters/FiltersContext/ContextFilters';
 
 const Admin = () => {
     return (
         <>
-            <LoginContainer />
-            <FiltersContainer />
+            {/* <LoginContainer /> */}
+            <ContextOfFilters>
+                <FiltersContainer />
+            </ContextOfFilters>
         </>
     );
 };
