+++ packages/admin-panel/src/components/Filters/FiltersContainer.js
@@ -1,82 +1,11 @@
-import React, { useState } from 'react';
+import React, { useContext } from 'react';
 import { CardContainer } from './CardContainer';
 import "./filters.css";
+import { ContextFilters } from './FiltersContext/ContextFilters';
 
 export const FiltersContainer = () => {
 
-    const [checked, setChecked] = useState({
-        checkedOne: false,
-        checkedTwo: false,
-        checkedThree: false,
-        checkedFour: false,
-        checkedFive: false,
-        checkedSixx: false
-    })
-
-    const [show, setShow] = useState({
-        showOne: false,
-        showTwo: false,
-        showThree: false,
-    })
-
-    const [openFilters, setOpenFilters] = useState(false)
-
-    const handleSetShow = (state, value) => {
-        setShow({
-            ...checked,
-            [state]: value,
-        })
-    }
-
-    const handleSetChecked = (state, value) => {
-        setChecked({
-            ...checked,
-            [state]: value,
-        })
-    }
-
-    const handleSetDelete = (stateOne, stateTwo, valueOne, valueTwo) => {
-        setChecked({
-            ...checked,
-            [stateOne]: valueOne,
-            [stateTwo]: valueTwo
-        })
-    }
-
-    const users = [
-        {
-            nombre: "user apellido apellido",
-            tipo: "huesped",
-            estado: "activo",
-            validacion: "si",
-        },
-        {
-            nombre: "user apellido apellido",
-            tipo: "huesped",
-            estado: "activo",
-            validacion: "si",
-        },
-        {
-            nombre: "user apellido apellido",
-            tipo: "anfitrion",
-            estado: "inactivo",
-            validacion: "si",
-        },
-        {
-            nombre: "user apellido apellido",
-            tipo: "anfitrion",
-            estado: "inactivo",
-            validacion: "si",
-        },
-        {
-            nombre: "user apellido apellido",
-            tipo: "anfitrion",
-            estado: "activo",
-            validacion: "no",
-        }
-    ]
-
-    const filterUsers = users.filter((u) => checked.checkedOne && u.tipo === "anfitrion" || checked.checkedTwo && u.tipo === "huesped" || checked.checkedThree && u.estado === "activo" || checked.checkedFour && u.estado === "inactivo" || checked.checkedFive && u.validacion === "si" || checked.checkedSixx && u.validacion === "no")
+    const { checked, setChecked, show, openFilters, setOpenFilters, handleSetShow, handleSetChecked, handleSetDelete, filterUsers, users } = useContext(ContextFilters)
 
     return (
         <>
@@ -122,13 +51,13 @@ export const FiltersContainer = () => {
                             <CardContainer type={"Estado"} titleOne="Activo" titleTwo="Inactivo" descriptionOne="Descripcion para activo" descriptionTwo="Descripcion para inactivo" show={show.showTwo} checkedOne={checked.checkedThree} checkedTwo={checked.checkedFour} handleSetShow={() => handleSetShow("showTwo", !show.showTwo)} handleChexboxOne={(e) => handleSetChecked("checkedThree", e.target.checked)} handleChexboxTwo={(e) => handleSetChecked("checkedFour", e.target.checked)} handleChexboxOneDelete={(e) => handleSetDelete("checkedThree", "checkedFour", false, false)} />
                             <CardContainer type={"Validación"} titleOne="Sí" titleTwo="No" descriptionOne="Descripcion para sí" descriptionTwo="Descripcion para no" show={show.showThree} checkedOne={checked.checkedFive} checkedTwo={checked.checkedSixx} handleSetShow={() => handleSetShow("showThree", !show.showThree)} handleChexboxOne={(e) => handleSetChecked("checkedFive", e.target.checked)} handleChexboxTwo={(e) => handleSetChecked("checkedSixx", e.target.checked)} handleChexboxOneDelete={(e) => handleSetDelete("checkedFive", "checkedSixx", false, false)} />
                             <div className='containerButtons'>
-                            <button onClick={() => setChecked({checkedOne: false, checkedTwo: false, checkedThree: false, checkedFour: false, checkedFive: false,checkedSixx: false})} className='buttonDelete'>
-                                Borrar
-                            </button>
-                            <button className='buttonSave'>
-                                Guardar
-                            </button>
-                        </div>
+                                <button onClick={() => setChecked({ checkedOne: false, checkedTwo: false, checkedThree: false, checkedFour: false, checkedFive: false, checkedSixx: false })} className='buttonDelete'>
+                                    Borrar
+                                </button>
+                                <button className='buttonSave'>
+                                    Guardar
+                                </button>
+                            </div>
                         </div>
                     </div>
                 </div>
