+++ packages/admin-panel/src/components/Filters/filters.css
@@ -40,7 +40,7 @@
     flex-direction: column;
     border-radius: 32px;
     box-shadow: 0px 4px 24px rgba(0, 0, 0, 0.2);
-    height: 215px;
+    height: 226px;
     justify-content: space-between;
     padding: 1rem;
     position: absolute;
