+++ packages/shared/src/index.js
@@ -2,7 +2,6 @@ export { Button } from './components/Button';
 export { Places, Perks, HouseRules} from '@booking-ui/places';
 export { Geolocation } from '@booking-ui/geolocation';
 export { MessagingSystem, TextBox } from '@booking-ui/messaging-system';
-export { Demo } from '@booking-ui/search-engine';
 export { UserAccount } from '@booking-ui/user-account';
 export { Demo, ResultsPage } from '@booking-ui/search-engine';
 export * from '@booking-ui/authentication';
